Add properties-minimal context to avoid 25000 context sanity limit
Sublime added a limit to the number of contexts that can be on the stack at one
time. The assumption was that more than 25000 contexts must be an include loop,
but CSS3 is such a beast that it actually hits this limit, without any include
loop, in the properties context. The bug was caused by the HTML syntax using
the rule-list-body context, which pushed the beastly properties context.

My hacky solution is to make a second properties context, with ~150 or so
properties removed. I couldn't find any data about which properties were the
least commonly used, so I just guessed. Undoubtedly, I cut out a property I
shouldn't have, but at least this fixes the bug. The properties will only be
missed when editing CSS inside a <style> in HTML.