Remove @apply. Fix custom properties.
The @apply spec has been abandoned and will never become a part of
official CSS. I initially added support for @apply to support cssnext.
The spec author explains the decision here:
    https://www.xanthir.com/b4o00

A happy coincidence is that this fixes highlighting values of custom
selectors. A bug in the regex related to supporting @apply was preventing
the values of customer properties like --foo: from being highlighted.
See this issue for details:
    https://github.com/y0ssar1an/CSS3/issues/112