Switch from CircleCI to GitHub Actions
Syntax tests are currently broken on CircleCI. The latest docs for the
`sublimetext/unittesting` image provide a GitHub Actions config. This
seems to be the preferred way to run the `unittesting` image.

Unfortunately, the nesting syntax test is extremely broken and I don't
know how to fix it. So that set of tests has been deleted.